Изменил структуру проекта. Положил все схемы и сервисы к их модулям

+++ app/products/router.py
@@ -1,19 +1,22 @@
-from fastapi import APIRouter, HTTPException, UploadFile, Form, Path, Depends, File
+import os
+
+from fastapi import APIRouter, HTTPException, BackgroundTasks, UploadFile, Path, Depends, File
 from sqlalchemy.orm import Session
 
 from core import get_db, get_current_user, save_product_image, delete_image
 from models import User, Product
-from schemas.product import CreatingProductDTO, CreatingProductResponse, PatchingProductDTO, GettingProductResponse
+from .schemas import CreatingProductDTO, CreatingProductResponse, PatchingProductDTO, GettingProductResponse
 
 router = APIRouter(prefix="/products", tags=["Product"])
 
 @router.post('', status_code=201, response_model=CreatingProductResponse)
-async def create_product(product_data: CreatingProductDTO = Depends(),
-                         image: UploadFile | None = File(None),
+async def create_product(background_tasks: BackgroundTasks,
+                         product_data: CreatingProductDTO = Depends(),
+                         image: UploadFile | None = File(None, max_length=15 *1024*1024, media_type=['image/png', 'image/jpeg']),
                          seller: User = Depends(get_current_user),
                          db: Session = Depends(get_db)
                          ):
-    
+    # ИСПРАВИТЬ КОД ДЛЯ ФОТОК
     product = Product(name=product_data.name,
                       price=product_data.price,
                       description=product_data.description,
@@ -21,10 +24,12 @@ async def create_product(product_data: CreatingProductDTO = Depends(),
     db.add(product)
     db.flush()
 
-    img = await save_product_image(image, product.id)
-    product.image = img
-    
+    #img_path = os.path.join("media", "product", str(product.id))
+    #ext = image.filename.split('.')[-1]
+    #product.image = f"{img_path}.{ext}"     
     db.commit()
+
+    #background_tasks.add_task(save_product_image(image, product.id))
 
     return {"status": "created", "product_id": product.id}
 
@@ -41,7 +46,7 @@ def get_product(product_id: int = Path(..., ge=1),
 @router.patch('/{product_id}', status_code=204)
 async def change_product(product_id: int = Path(..., ge=1),
                          product_data: PatchingProductDTO = Depends(),
-                         image: UploadFile | None = File(None),
+                         image: UploadFile | None = File(None, max_length=15 *1024*1024, media_type=['image/png', 'image/jpeg']),
                          seller: User = Depends(get_current_user),
                          db: Session = Depends(get_db)
                          ):
@@ -70,8 +75,8 @@ async def delete_product(product_id: int = Path(..., ge=1),
     if not product:
         raise HTTPException(status_code=404, detail="You cannot delete this product")
     
-    try:
+    if product.image:
         await delete_image("product", product.image)
-    finally:
-        db.delete(product)
-        db.commit()
+        
+    db.delete(product)
+    db.commit()
